get the c++ AtomPairs and torsion stuff working on windows. This builds under vc71, but still hasn't been set up under vc80.

+++ Code/GraphMol/Descriptors/test.cpp
@@ -269,8 +269,8 @@ void testTorsions(){
 
   ROMol *mol;
   SparseIntVect<long long int> *fp;
-  unsigned int tgt;
-  unsigned int c1,c2,c3,c4;
+  unsigned long long int tgt;
+  unsigned long long int c1,c2,c3,c4;
   std::vector<unsigned int> codes;
 
   mol = SmilesToMol("CCCC");
